Revise input buffer size checks
* src/cpiohdr.h: Define SIZE_MAX unless already defined.
* src/extern.h (io_block_size): Change type to size_t.
* src/global.c: Likewise.
* src/main.c: Use DISK_IO_BLOCK_SIZE instead of 512.
(doc): Remove unnecessary quoting of the program name.
(get_block_size): Operate on size_t.
(copyin_buf_size): Remove unnecessary check.  Check for overflow instead.
* src/util.c: Use DISK_IO_BLOCK_SIZE instead of hardcoded 512.
* tests/sizemax.c: New auxiliary program.
* tests/Makefile.am: Build sizemax.
* tests/big-block-size.at: Rewrite.

+++ src/extern.h
@@ -28,7 +28,7 @@ enum archive_format
 
 extern enum archive_format archive_format;
 extern int reset_time_flag;
-extern int io_block_size;
+extern size_t io_block_size;
 extern int create_dir_flag;
 extern int rename_flag;
 extern char *rename_batch_file;
